refactor: new directory pages
