+++ src/client/pages/Favourites.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useSession } from "../context/SessionContext";
 import { useNavigate } from "react-router-dom";
-import PictureGrid from "./PictureGrid";
+import PictureGrid from "../components/PictureGrid";
 
 export default function Favourites() {
   const { session } = useSession();
